Move register code to headers
