+++ src/module.cpp
@@ -6,17 +6,17 @@
 #include <boost/histogram/python/pybind11.hpp>
 
 void register_version(py::module &);
-py::module register_storage(py::module &);
-py::module register_axis(py::module &);
-py::module register_histogram(py::module &);
+py::module register_storages(py::module &);
+py::module register_axes(py::module &);
+py::module register_histograms(py::module &);
 void register_make_histogram(py::module &, py::module &);
 py::module register_accumulators(py::module &);
 
 PYBIND11_MODULE(histogram, m) {
     register_version(m);
-    register_storage(m);
-    register_axis(m);
-    py::module hist = register_histogram(m);
+    register_storages(m);
+    register_axes(m);
+    py::module hist = register_histograms(m);
     register_make_histogram(m, hist);
     register_accumulators(m);
 }
